# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Share Link API",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateSHLink": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "A string representing the share link's unique user identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "name": {
            "type": "string",
            "description": "A string representing the share link's name.",
            "example": "Jane Doe's Shlink"
          },
          "configPasscode": {
            "type": "string",
            "description": "An optional string representing the share link's password.",
            "example": null
          },
          "configExp": {
            "type": "datetime",
            "description": "An optional date representing the date when the share link expires.",
            "example": null
          }
        }
      },
      "SHLinkFile": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "description": "A string representing the content type of share link file.",
            "example": "application/smart-api-access"
          },
          "embedded": {
            "type": "string",
            "description": "An optional string representing the content of a share link's file.",
            "example": null
          },
          "location": {
            "type": "string",
            "description": "A string representing the share link's API link.",
            "example": "https://example.com/api/v1/shlinks/shlink-id/endpoints/endpoint-id?ticket=unique-ticket"
          }
        }
      },
      "SHLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share links unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "userId": {
            "type": "string",
            "description": "A string representing the share link's unique user identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "name": {
            "type": "string",
            "description": "A string representing the share link's name.",
            "example": "Jane Doe's Shlink"
          },
          "configPasscode": {
            "type": "string",
            "description": "An optional string representing the share link's password.",
            "example": null
          },
          "configExp": {
            "type": "datetime",
            "description": "An optional date representing the date when the share link expires.",
            "example": null
          },
          "passcodeFailureRemaining": {
            "type": "number",
            "description": "An integer representing how many passcode failure attempts are remaining for the share link.",
            "example": 5
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's unique management token used for managing the share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          }
        }
      },
      "SHLinkMini": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share links unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "name": {
            "type": "string",
            "description": "A string representing the share link's name.",
            "example": "Jane Doe's Shlink"
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's unique management token used for managing the share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "url": {
            "type": "string",
            "description": "A string representing the share link's encoded url.",
            "example": "5AMl62z2XDmgrh2XsI2O"
          },
          "passwordRequired": {
            "type": "boolean",
            "description": "A boolean indicating whether the share link has a passcode or not.",
            "example": true
          },
          "expiryDate": {
            "type": "datetime",
            "required": false,
            "description": "An optional date representing when the share link expires.",
            "example": "2024-09-09T09:45:00.479+00:00"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SHLinkFile"
            }
          }
        }
      },
      "SHLinkRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share links unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's unique management token used for managing the share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "recipient": {
            "type": "string",
            "description": "A string representing the share link's recipient.",
            "example": "Jane Doe"
          }
        }
      },
      "SHLinkUpdate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share links unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "passcode": {
            "type": "string",
            "description": "An optional string representing the share link's new passcode.",
            "example": null
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's unique management token used for managing the share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "oldPasscode": {
            "type": "string",
            "description": "An optional string representing the share link's old passcode.",
            "example": null
          },
          "expiryDate": {
            "type": "datetime",
            "required": false,
            "description": "An optional date representing when the share link expires.",
            "example": "2024-09-09T09:45:00.479+00:00"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The user unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "patientId": {
            "type": "string",
            "description": "The patient's unique identifier in the external system.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "userId": {
            "type": "string",
            "description": "The user's unique identifier in the SSO system.",
            "example": "5AMl62z2XDmgrh2XsI2O"
          }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/v1/shlinks/{id}": {
      "post": {
        "tags": [
          "Share Links"
        ],
        "description": "Get a share link.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/SHLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get Share Link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/SHLinkMini"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Share Links"
        ],
        "description": "Update A Share link.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/SHLinkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Share Link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/SHLinkMini"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks": {
      "post": {
        "tags": [
          "Share Links"
        ],
        "description": "Create a server config.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/CreateSHLink"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create Server Config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/SHLink"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Share Links"
        ],
        "description": "Get share links.",
        "responses": {
          "200": {
            "description": "Gets all the signed in user's share links.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SHLinkMini"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}/ips": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get a user's patient summary data.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the user's unique identifier.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "International Patient Summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get a user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the user's unique identifier.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}